pyproject.toml
[tool.poetry]
name = "ml-serv"
version = "0.1.0"
description = ""
authors = ["Ilya <59098537+AzekAriman@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.26.0"}
fastapi-security = "^0.5.0"
sqlalchemy = "^2.0.25"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
